Move acceptance test to acceptance module
- Removed an unused file
- Code cleanups
